rag-fastapi/requirements.txt
fastapi
chromadb
openai
python-dotenv
uvicorn
